Seperate numpy and xarray interfaces
The import in the SAM interface was broken. I took the opporuntity to move the
numpy specific code to its own module.
