+++ uwnet/xarray_interface.py
@@ -68,21 +68,6 @@ def call_with_xr(self, ds, **kwargs):
     return _torch_dict_to_dataset(output, ds.coords)
 
 
-def _numpy_dict_to_torch_dict(inputs):
-    return {key: torch.from_numpy(val) for key, val in inputs.items()}
-
-
-def _torch_dict_to_numpy_dict(output):
-    return {key: val.numpy() for key, val in output.items()}
-
-
-def call_with_numpy_dict(self, inputs, **kwargs):
-    """Call the neural network with numpy inputs"""
-    with torch.no_grad():
-        return pipe(inputs, _numpy_dict_to_torch_dict,
-                    partial(self, **kwargs), _torch_dict_to_numpy_dict)
-
-
 class XRCallMixin(object):
     """PyTorch module for predicting Q1, Q2 and maybe Q3"""
     call_with_xr = call_with_xr
